refactored removing BigDecimal

+++ src/test/java/com/stark/unifi/datanalyze/analyzer/TextAnalyzerTest.java
@@ -2,11 +2,8 @@ package com.stark.unifi.datanalyze.analyzer;
 
 import static org.junit.Assert.assertEquals;
 
-import java.math.BigDecimal;
-
 import org.junit.Test;
 
-import com.stark.unifi.datanalyze.analyzer.TextAnalyzer;
 import com.stark.unifi.datanalyze.model.Document;
 import com.stark.unifi.datanalyze.util.FileContentReader;
 
@@ -20,12 +17,12 @@ public class TextAnalyzerTest {
 		String text = new FileContentReader().read("/txt/basic.txt");
 		Document result = analyzer.analyzeText(text);
 		
-		assertEquals(BigDecimal.valueOf(6), result.getSentenceCount());
-		assertEquals(BigDecimal.valueOf(36), result.getWordCount());
-		assertEquals(BigDecimal.valueOf(205), result.getCharacterCount());
-		assertEquals(BigDecimal.valueOf(137), result.getSyllableCount());
-		assertEquals(BigDecimal.valueOf(17), result.getComplexWordCount(3));
-		assertEquals(BigDecimal.valueOf(12), result.getLongWordCount());
+		assertEquals(6, result.getSentenceCount());
+		assertEquals(36, result.getWordCount());
+		assertEquals(205, result.getCharacterCount());
+		assertEquals(137, result.getSyllableCount());
+		assertEquals(17, result.getComplexWordCount(3));
+		assertEquals(12, result.getLongWordCount());
 	}
 	
 }
